v1_deployment/requirements.txt
streamlit
torch
chromadb
rank_bm25
open_clip_torch
ollama
Pillow
